Module wlc::input::keyboard

Keyboard related functionality

Modules

Keysyms

Structs

Keymap
Compiled keymap object.

State
Keyboard state object.

Enums

Key

Functions

current_keys
Get currently held keys

keysym_for_key
Utility function to convert raw keycode to keysym. Passed modifiers may transform the key

utf32_for_key
Utility function to convert raw keycode to Unicode/UTF-32 codepoint. Passed modifiers may transform the key

xkb_keymap
Internal xkb_keymap exposed. You can use it to do more advanced key handling

xkb_state
Internal xkb_state exposed. You can use it to do more advanced key handling. However you should avoid messing with its state

Type Definitions

Keysym
A number used to represent the symbols generated from a key on a keyboard.