Module wlc::render

wlc's Rendering extensions

Enabled by feature render

Some of the original function may be added as functions to View and Output, when this feature is enabled

The types/functions in this module provide some basic rendering capabilities.

For more advanced drawing you should directly use GLES2. This is not documented as it's currently relying on the implementation details of wlc.

Structs

GLES2Renderer
OpenGL ES 2 Renderer

NoPixelFormat
PixelFormat of the NoRenderer

NoRenderer
Headless renderer

NoTexture
Texture of the NoRenderer

RenderOutput
Output with Render Extensions

RenderView
View with Render Extensions

Enums

GLES2PixelFormat
Supported PixelFormats of the GLES2Renderer

RenderInstance
Wlc's Renderer Types

SurfaceFormat
Possible Surface formats

Traits

PixelFormat
Renderer specific pixel format marker

Renderer
Renderer API

Texture
Renderer specific Texture marker

Type Definitions

GLTexture
Texture of the GLES2Renderer (Raw OpenGL ES 2 Texture)