Trait wlc::render::Renderer

pub trait Renderer {
    type PixelFormat: PixelFormat;
    type Texture: Texture;
    fn pixels_write(&mut self,
                    format: Self::PixelFormat,
                    geometry: Geometry,
                    data: &[u8]);
    fn pixels_read(&self,
                   format: Self::PixelFormat,
                   geometry: &mut Geometry)
                   -> Vec<u8>;
}

Renderer API

Associated Types

type PixelFormat: PixelFormat

PixelFormat understood by this Renderer

type Texture: Texture

Texture understood by this Renderer

Required Methods

fn pixels_write(&mut self,
format: Self::PixelFormat,
geometry: Geometry,
data: &[u8])

Write pixel data with the specific format to output's framebuffer.

If the geometry is out of bounds, it will be automaticall clamped.

fn pixels_read(&self,
format: Self::PixelFormat,
geometry: &mut Geometry)
-> Vec<u8>

Read pixel data from output's framebuffer.

If the geometry is out of bounds, it will be automatically clamped. Potentially clamped geometry will change geometry, to indicate width / height of the returned data.