Struct wlc::View

#[repr(C)]

pub struct View;

A View managed by Wlc

Methods

impl View

fn set_no_focus()

Set no view focused

Safety

Dont call this function on another thread, then the main thread

fn focus(&self)

Focus view

fn close(&self)

Close view

fn output(&self) -> &Output

Get current Output

fn set_output(&self, output: &Output)

Set Output

fn send_to_back(&self)

Send behind everything else

fn send_below(&self, other: &View)

Send below another view

fn bring_to_front(&self)

Send above another view

fn bring_above(&self, other: &View)

Bring to front of everything

fn visibility(&self) -> Flags

Get current visibility bitmask

fn set_visibility(&self, visibility: Flags)

Set visibility bitmask

fn geometry(&self) -> Geometry

Get current geometry. (what client sees)

fn visible_geometry(&self) -> Geometry

Get visible geometry. (what wlc displays)

fn set_geometry(&self, edge: Flags, geometry: Geometry)

Set geometry. Set edges if the geometry change is caused by interactive resize

fn view_type(&self) -> Flags

Get type bitfield

fn set_view_type(&self, view_type: Flags, toggle: bool)

Set type bit. Toggle indicates whether it is set or not

fn state(&self) -> Flags

Get current state bitfield

fn set_state(&self, state: Flags, toggle: bool)

Set state bit. Toggle indicates whether it is set or not

fn parent(&self) -> Option<&View>

Get parent view

fn set_parent(&self, parent: &View)

Set parent view

fn title(&self) -> Cow<str>

Get title

fn class(&self) -> Cow<str>

Get class. (shell-surface only)

fn instance(&self) -> Cow<str>

Get instance. (shell-surface only)

fn app_id(&self) -> Cow<str>

Get app id. (xdg-surface only)

fn pid(&self) -> pid_t

Get pid

fn positioner(&self) -> Option<&Positioner>

Get the positioner of the View if one exists

The Positioner provides a collection of rules for the placement of a child View relative to it's parent. Compositors are expected to value these hints, but technically not required.

fn weak_reference(&self) -> WeakView

Get a weak reference of the View that may outlive its referenced view

Since View is always limited in its use by its lifetime, it is not very suitable for storing. This function allows you to optain a weak reference, that may outlive the view it is referencing.