Struct wlc::WeakView

pub struct WeakView(_, _);

Weak reference to a view

Can be optained by view.weak_reference()

Methods

impl WeakView

unsafe fn upgrade(&self) -> Option<&View>

Upgrade your weak reference to an actual View

Safety

This function is unsafe, because it creates a lifetime bound to the WeakView, which may live forever.. But no view lives forever and might be destroyed by its process at any time A disconnection is signaled by Callback::view_destroyed and the View is deallocated shortly after. Because of this using this function one may create an invalid View reference. Also using this function it is possible to optain multiple mutable references, which is also unsafe

See WeakView::run for a safe variant

fn run<F, R>(&self, runner: F) -> Option<R> where F: FnOnce(&View) -> R

Run a function on the referenced View, if it still exists

Returns the result of the function, if successful

Safety

By enforcing a rather harsh limit on the lifetime of the view to a short lived scope of an anonymous function, this function makes sure the view does not live longer then it exists.