Module wlc::wayland

wlc's Wayland extensions

Enabled by feature wayland

Some of the original function may be directly added as functions to View and Output, when this feature is enabled.

Modules

sys
The following wlc function are exposed here for own implementation as a general purpose implementation is out of scope of this crate.

Structs

WlcSubSurface
Subsurface of an WlcSurface

WlcSurface
Internal Surface Struct

Functions

output_from_wl_surface
Returns a weak output handle from WlOutput resource

view_from_wl_surface
Returns a weak view handle from WlSurface resource

wlc_resource_from_wl_surface
Returns internal wlc surface from WlSurface resource