Struct wlc::wayland::WlcSurface

#[repr(C)]

pub struct WlcSurface;

Internal Surface Struct

Methods

impl WlcSurface

fn size(&self) -> Size

Get surface size

fn wl_surface(&self) -> WlSurface

Return wl_surface resource from internal wlc surface

fn sub_surfaces(&self) -> &[&WlcSubSurface]

Returns a list of the subsurfaces of the given surface

fn flush_frame_callbacks(&self)

Adds frame callbacks of the given surface for the next output frame. It applies recursively to all subsurfaces.

Useful when the compositor creates custom animations which require disabling internal rendering, but still need to update the surface textures (for ex. video players).