Module wlc::wayland::sys

The following wlc function are exposed here for own implementation as a general purpose implementation is out of scope of this crate.

Functions

wlc_get_wl_display⚠
Returns Wayland display.

wlc_view_from_surface⚠
Turns wl_surface into a wlc view. Returns 0 on failure. This will also trigger view.created callback as any view would. For the extra arguments see details of wl_resource_create and wl_resource_set_implementation. The extra arguments may be set NULL, if you are not implementing Wayland interface for the surface role.

wlc_view_get_role⚠
Returns surface role resource from view handle. Return value will be NULL if the view was not assigned role or created with wlc_view_create_from_surface().